store W.h in DB for more accuracy. openDCIM still uses kW.h.

+++ db-4.1-to-PUE_DCEM.sql
@@ -91,7 +91,7 @@ CREATE TABLE IF NOT EXISTS fac_ElectricalMeasure (
   Wattage1 INT(11) NOT NULL,
   Wattage2 INT(11) NOT NULL,
   Wattage3 INT(11) NOT NULL,
-  Energy INT(11) NOT NULL,
+  Energy BIGINT(20) NOT NULL,
   Date DATETIME NOT NULL,
   KEY MPID (MPID),
   UNIQUE KEY (MPID, Date)
